feat: introduce HasComponentsOfType for typed child containers
Adds HasComponentsOfType<T extends Component> as the typed counterpart of
HasComponents, for containers that should only accept a specific child
component type (e.g. a breadcrumb trail that only accepts breadcrumb items).
HasComponents now extends HasComponentsOfType<Component> and keeps the
untyped add(String) convenience. All existing default methods are preserved
via inheritance.
